+++ src/AM.cpp
@@ -365,6 +365,8 @@ int main(int argc, char **argv)
 		return 1;
 	}
 
+	Ambition::Control::useGraphicalErrorReporting();
+
 	err.set_extra_handler( extra_error_handler );   // set extra error handler, save the game when a error happens
 
 	Ambition::Control::displayNews();