+++ src/ambition/7kaaInterface/control.cc
@@ -291,6 +291,11 @@ void unlockBuffer(
   Ambition::unlockBuffer(buffer);
 }
 
+void useGraphicalErrorReporting(
+) {
+  Ambition::Control::useGraphicalErrorReporting();
+}
+
 
 /* Private functions */